app.py
import streamlit as st
import cv2
import numpy as np
from PIL import Image
import torch

# Set page title and icon
st.set_page_config(page_title="Pothole Detection AI", page_icon="🕳️")

# Title and description
st.title("🕳️ Pothole Detection AI")
st.write("Upload an image to detect potholes using our AI model")

# File uploader
uploaded_file = st.file_uploader("Choose an image...", type=['jpg', 'jpeg', 'png'])

if uploaded_file is not None:
    # Read the image
    image = Image.open(uploaded_file)
    st.image(image, caption='Uploaded Image', use_column_width=True)
    
    # Add your model loading and prediction code here
    # You'll need to load your best.pt model and run inference
    
    st.success("Pothole detection completed!")
    # Display results here